helpers: split the code in modules
This change moves the code to dedicated modules.
